# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Resume Customizer API
  description: |
    Resume Customizer exposes a REST API for triggering the resume-tailoring pipeline, tracking runs,
    and retrieving artifacts (including the generated LaTeX resume).

    The endpoints and their descriptions are based on the repository README's REST API section.
  version: "0.1.0"
servers:
  - url: http://localhost:8080/v1
    description: Local development (v1 API)
  - url: http://localhost:8080
    description: Legacy endpoints (deprecated)
...
paths:
...
  /v1/auth/register:
    post:
      tags: [authentication]
      summary: Register a new user
      description: Creates a new user account with email and password
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/login:
    post:
      tags: [authentication]
      summary: Login user
      description: Authenticates a user and returns a JWT token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /run:
    post:
      tags: [pipeline]
      summary: Start pipeline (background)
      description: Starts the pipeline asynchronously and returns a run ID immediately.
      operationId: startRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCreateRequest"
            examples:
              basic:
                summary: Start a run from a job URL
                value:
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  job_url: "https://www.linkedin.com/jobs/view/123456789"
      responses:
        "202":
          description: Accepted (run started)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /run/stream:
    post:
      tags: [pipeline]
      summary: Start pipeline (SSE)
      description: Starts the pipeline and streams progress events using Server-Sent Events (SSE).
      operationId: startRunStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCreateRequest"
      responses:
        "200":
          description: |
            SSE stream of progress events. The first event will always be `run_started`
            containing the `run_id` for tracking. Subsequent events are `step` events
            for pipeline progress, followed by `complete` or `error` at the end.
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SseEvent"
              examples:
                run_started:
                  summary: First event with run_id for tracking
                  value: |
                    event: step
                    data: {"step":"run_started","category":"lifecycle","message":"Pipeline run started","run_id":"550e8400-e29b-41d4-a716-446655440000"}
                progress:
                  summary: Subsequent pipeline progress events
                  value: |
                    event: step
                    data: {"step":"job_profile","category":"ingestion","message":"Parsed job profile: Software Engineer at Acme Corp"}
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/runs:
    get:
      tags: [runs]
      summary: List all runs
      description: Lists pipeline runs with optional filters.
      operationId: listRuns
      parameters:
        - in: query
          name: company
          schema: { type: string }
          description: Filter by company name
        - in: query
          name: status
          schema:
            type: string
            enum: [queued, running, completed, failed, canceled]
          description: Filter by run status
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
          description: Max number of runs to return
      responses:
        "200":
          description: A list of runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/status/{id}:
    get:
      tags: [runs]
      summary: Get run status
      description: Returns status, company, role (and other metadata) for a run.
      operationId: getRunStatus
      parameters:
        - $ref: "#/components/parameters/RunIdPath"
      responses:
        "200":
          description: Run status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunStatus"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/runs/{id}:
    delete:
      tags: [runs]
      summary: Delete run
      description: Deletes a run and cascades deletion to associated artifacts.
      operationId: deleteRun
      parameters:
        - $ref: "#/components/parameters/RunIdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/artifacts:
    get:
      tags: [artifacts]
      summary: List artifacts
      description: Lists artifacts with optional filters.
      operationId: listArtifacts
      parameters:
        - in: query
          name: run_id
          schema: { type: string, format: uuid }
          description: Filter by run ID
        - in: query
          name: step
          schema: { type: string }
          description: Filter by pipeline step
        - in: query
          name: category
          schema: { type: string }
          description: Filter by artifact category
      responses:
        "200":
          description: A list of artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtifactSummary"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/artifact/{id}:
    get:
      tags: [artifacts]
      summary: Get artifact by ID
      description: Returns full artifact JSON content.
      operationId: getArtifact
      parameters:
        - $ref: "#/components/parameters/ArtifactIdPath"
      responses:
        "200":
          description: Full artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/runs/{id}/artifacts:
    get:
      tags: [artifacts]
      summary: List artifacts for a run
      description: Lists artifacts for a single run (summary view).
      operationId: listRunArtifacts
      parameters:
        - $ref: "#/components/parameters/RunIdPath"
      responses:
        "200":
          description: A list of artifacts for the run
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtifactSummary"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/runs/{id}/resume.tex:
    get:
      tags: [artifacts]
      summary: Download LaTeX resume
      description: Downloads the generated LaTeX resume for a run as plain text.
      operationId: downloadResumeTex
      parameters:
        - $ref: "#/components/parameters/RunIdPath"
      responses:
        "200":
          description: LaTeX document
          content:
            text/plain:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/users:
    post:
      tags: [users]
      summary: Create user
      description: Creates a user profile.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/users/{id}:
    get:
      tags: [users]
      summary: Get user
      description: Returns a user profile.
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [users]
      summary: Update user
      description: Updates a user profile.
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/users/{id}/password:
    put:
      tags: [authentication]
      summary: Update user password
      description: Updates the password for a specific user. The authenticated user must match the user ID in the path.
      operationId: updateUserPassword
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordRequest"
      responses:
        "200":
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password updated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Unauthorized (invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden (cannot update another user's password)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/users/{id}/jobs:
    get:
      tags: [jobs]
      summary: List jobs
      description: Lists employment history for a user.
      operationId: listJobs
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [jobs]
      summary: Create job
      description: Creates a job (employment entry) for a user.
      operationId: createJob
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreateRequest"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/jobs/{id}:
    put:
      tags: [jobs]
      summary: Update job
      description: Updates a job.
      operationId: updateJob
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdateRequest"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [jobs]
      summary: Delete job
      description: Deletes a job and its experiences.
      operationId: deleteJob
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/jobs/{id}/experiences:
    get:
      tags: [experiences]
      summary: List experiences
      description: Lists bullet-point experiences for a job.
      operationId: listExperiences
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      responses:
        "200":
          description: Experiences
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Experience" }
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [experiences]
      summary: Add experience
      description: Adds an experience bullet to a job.
      operationId: createExperience
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceCreateRequest"
      responses:
        "201":
          description: Created experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/experiences/{id}:
    put:
      tags: [experiences]
      summary: Update experience
      description: Updates an experience bullet.
      operationId: updateExperience
      parameters:
        - $ref: "#/components/parameters/ExperienceIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceUpdateRequest"
      responses:
        "200":
          description: Updated experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [experiences]
      summary: Delete experience
      description: Deletes an experience bullet.
      operationId: deleteExperience
      parameters:
        - $ref: "#/components/parameters/ExperienceIdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/users/{id}/education:
    get:
      tags: [education]
      summary: List education
      description: Lists education history for a user.
      operationId: listEducation
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: Education entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Education" }
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [education]
      summary: Add education
      description: Adds an education entry for a user.
      operationId: createEducation
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationCreateRequest"
      responses:
        "201":
          description: Created education entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Education"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/users/{id}/experience-bank:
    get:
      tags: [users]
      summary: Export experience bank
      description: Exports the user's experience bank JSON in a pipeline-compatible format.
      operationId: exportExperienceBank
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: Experience bank export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperienceBankExport"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /v1/auth/login or /v1/auth/register
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            bad_request:
              value:
                error: bad_request
                message: Invalid request payload
...
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          description: User's full name
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          format: password
          minLength: 8
          description: User's password (minimum 8 characters)
        phone:
          type: string
          nullable: true
          description: User's phone number (optional)
...
    LoginResponse:
      type: object
      required:
        - user
        - token
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: JWT token for authenticated requests
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."